Add tab packages to project
