+++ app/src/main/java/com/example/referenciapp/screens/exercise/ExerciseFragment.kt
@@ -1,12 +1,12 @@
-package com.example.referenciapp
+package com.example.referenciapp.screens.exercise
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.TextView
 import androidx.databinding.DataBindingUtil
+import com.example.referenciapp.R
 import com.example.referenciapp.databinding.FragmentExerciseBinding
 
 /**
