+++ app/src/main/java/com/example/referenciapp/screens/referenceMenu/ReferenceMenuFragment.kt
@@ -6,10 +6,10 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
-import com.example.referenciapp.DigitalTabFragment
-import com.example.referenciapp.PaperTabFragment
+import com.example.referenciapp.tabs.digitalTab.DigitalTabFragment
+import com.example.referenciapp.tabs.paperTab.PaperTabFragment
 import com.example.referenciapp.R
-import com.example.referenciapp.ViewPagerAdapter
+import com.example.referenciapp.tabs.ViewPagerAdapter
 import com.example.referenciapp.databinding.FragmentReferenceMenuBinding
 
 class ReferenceMenuFragment : Fragment() {
