+++ app/src/main/java/com/example/referenciapp/tabs/paperTab/PaperTabFragment.kt
@@ -1,5 +1,4 @@
-package com.example.referenciapp
-
+package com.example.referenciapp.tabs.paperTab
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
@@ -7,17 +6,19 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
 import androidx.navigation.Navigation
-import com.example.referenciapp.databinding.FragmentDigitalTabBinding
+import com.example.referenciapp.R
+import com.example.referenciapp.databinding.FragmentPaperTabBinding
 
-class DigitalTabFragment : Fragment() {
+class PaperTabFragment : Fragment() {
+
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-
-        val binding: FragmentDigitalTabBinding = DataBindingUtil.inflate(
+        // Inflate the layout for this fragment
+        val binding: FragmentPaperTabBinding = DataBindingUtil.inflate(
             inflater,
-            R.layout.fragment_digital_tab,
+            R.layout.fragment_paper_tab,
             container,
             false
         )
@@ -34,5 +35,5 @@ class DigitalTabFragment : Fragment() {
 
         return binding.root
     }
-}
 
+}
