+++ app/src/main/res/layout/fragment_paper_tab.xml
@@ -2,7 +2,7 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".PaperTabFragment">
+    tools:context=".tabs.paperTab.PaperTabFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
