+++ app/src/main/res/navigation/navigation.xml
@@ -27,7 +27,7 @@
     </fragment>
     <fragment
         android:id="@+id/exerciseFragment"
-        android:name="com.example.referenciapp.ExerciseFragment"
+        android:name="com.example.referenciapp.screens.exercise.ExerciseFragment"
         android:label="ExerciseFragment" />
     <fragment
         android:id="@+id/listMenuFragment"
